created own component for file upload
